+++ wrappers/table-bucket/versions.tf
@@ -4,7 +4,10 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 6.2"
+      version = ">= 6.42"
     }
+  }
+
+  provider_meta "aws" {
   }
 }